rdp: Remove the move pending list
weston_output_move is idempotent, so it doesn't matter if we move a head
that's already in the right place. So let's just move all the heads
instead of having a special move pending list at all.

This will make things easier later when the frontend and backend code do
different parts of monitor setup later.

Signed-off-by: Derek Foreman <derek.foreman@collabora.com>