rdp: rework head layout validation
Perform this on the compositor head_list instead of on the monitor array,
which is how it needs to work to be in the front end code, where it will
only have access to the compositor list, and only after the heads have
been created.

This also moves the validation code to after the heads have been created
and removed - it validates output configuration and scale parameters, but
wasn't really changing head allocation/deallocation anyway.

Signed-off-by: Derek Foreman <derek.foreman@collabora.com>